+++ app/Http/Controllers/TagsController.php
@@ -106,6 +106,12 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+
+        if ($tag->posts->count()>0){
+            session()->flash('error','Tag Cannot be Deleted, Because It Has Some Posts!');
+
+            return redirect()->back();
+        }
         $tag->delete();
 
         session()->flash('success','Tag Delete Successfully');
